Refactor: 커스텀 훅 및 컴포넌트 분리
