+++ src/app/page.jsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { PostList } from '@src/components/postList/PostList';
+import { PostList } from '@src/components/Postlist/PostList';
 
 export default function Home() {
   return (
