+++ src/app/posts/[[...userId]]/page.jsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { PostList } from '@src/components/postList/PostList';
+import { PostList } from '@src/components/Postlist/PostList';
 
 export default function Posts({ params }) {
   const blogUserId = params.userId ? params.userId[0] : '';
